+++ experimental/packages/opentelemetry-sdk-node/src/types.ts
@@ -15,14 +15,15 @@
  */
 
 import type { ContextManager, SpanAttributes } from '@opentelemetry/api';
-import { Sampler, TextMapPropagator } from '@opentelemetry/api';
+import { TextMapPropagator } from '@opentelemetry/api';
 import { InstrumentationOption } from '@opentelemetry/instrumentation';
 import { Resource } from '@opentelemetry/resources';
 import { MetricReader } from '@opentelemetry/sdk-metrics-base';
 import {
+  Sampler,
   SpanExporter,
   SpanLimits,
-  SpanProcessor
+  SpanProcessor,
 } from '@opentelemetry/sdk-trace-base';
 
 export interface NodeSDKConfiguration {