+++ packages/opentelemetry-sdk-trace-base/test/common/export/ConsoleSpanExporter.test.ts
@@ -18,10 +18,10 @@ import {
   SpanContext,
   TraceFlags,
 } from '@opentelemetry/api';
-import { AlwaysOnSampler } from '@opentelemetry/core';
 import * as assert from 'assert';
 import * as sinon from 'sinon';
 import {
+  AlwaysOnSampler,
   BasicTracerProvider,
   ConsoleSpanExporter,
   SimpleSpanProcessor,